+++ .github/workflows/test.yml
@@ -439,4 +439,4 @@ jobs:
     steps:
       - uses: actions/checkout@v4
       - uses: dtolnay/rust-toolchain@nightly
-      - run: cargo doc --workspace --exclude=gsym-in-apk --no-deps
+      - run: cargo doc --workspace --exclude=blazesym-dev --exclude=gsym-in-apk --no-deps
