+++ dev/Cargo.toml
@@ -39,5 +39,12 @@ xz2 = {version = "0.1.7", optional = true}
 zip = {version = "2.0.0", optional = true, default-features = false}
 zstd = {version = "0.13.1", default-features = false, optional = true}
 
+[dependencies]
+blazesym = {version = "=0.2.0-rc.2", path = "../"}
+libc = "0.2.137"
+
+[target.'cfg(target_os = "linux")'.dependencies]
+libbpf-rs = "0.24"
+
 [target.'cfg(any(target_os = "linux", target_os = "android"))'.build-dependencies]
 libbpf-sys = {version = "1.4.1", default-features = false, optional = true}
