+++ dev/build.rs
@@ -754,6 +754,12 @@ fn prepare_bench_files() {
 }
 
 fn main() {
+    let target_os = env::var("CARGO_CFG_TARGET_OS").unwrap();
+    if target_os == "linux" || target_os == "android" {
+        println!("cargo:rustc-cfg=linux");
+    }
+    println!("cargo:rustc-check-cfg=cfg(linux)");
+
     if cfg!(feature = "generate-unit-test-files")
         && !cfg!(feature = "dont-generate-unit-test-files")
     {
