+++ src/lib.rs
@@ -169,6 +169,7 @@ pub mod helper {
 pub mod __private {
     pub use crate::util::bytes_to_path;
     pub use crate::util::stat;
+    pub use crate::util::ReadRaw;
 }
