+++ tests/blazesym.rs
@@ -43,6 +43,9 @@ use blazesym::Pid;
 use blazesym::Result;
 use blazesym::SymType;
 
+#[cfg(linux)]
+use blazesym_dev::with_bpf_symbolization_target_addrs;
+
 use scopeguard::defer;
 
 use tempfile::tempdir;
@@ -803,6 +806,43 @@ fn symbolize_process_with_custom_dispatch() {
 
     test(process_dispatch);
     test(process_no_dispatch);
+}
+
+/// Test symbolization of a kernel address inside a BPF program.
+#[cfg(linux)]
+#[test]
+fn symbolize_kernel_bpf_program() {
+    with_bpf_symbolization_target_addrs(|handle_getpid, subprogram| {
+        let src = symbolize::Source::Kernel(symbolize::Kernel::default());
+        let symbolizer = Symbolizer::new();
+        let result = symbolizer
+            .symbolize(
+                &src,
+                symbolize::Input::AbsAddr(&[handle_getpid, subprogram]),
+            )
+            .unwrap();
+        let handle_getpid_sym = result[0].as_sym().unwrap();
+        assert_eq!(handle_getpid_sym.name, "handle__getpid");
+        let code_info = handle_getpid_sym.code_info.as_ref().unwrap();
+        assert_eq!(code_info.dir, None);
+        assert_eq!(
+            Path::new(&code_info.file).file_name(),
+            Some(OsStr::new("getpid.bpf.c"))
+        );
+        assert_eq!(code_info.line, Some(33));
+        assert_ne!(code_info.column, None);
+
+        let subprogram_sym = result[1].as_sym().unwrap();
+        assert_eq!(subprogram_sym.name, "subprogram");
+        let code_info = subprogram_sym.code_info.as_ref().unwrap();
+        assert_eq!(code_info.dir, None);
+        assert_eq!(
+            Path::new(&code_info.file).file_name(),
+            Some(OsStr::new("getpid.bpf.c"))
+        );
+        assert_eq!(code_info.line, Some(15));
+        assert_ne!(code_info.column, None);
+    })
 }
 
 /// Symbolize a normalized address from a binary with an artificially
